Fix xclLogMsg va_list processing -- cannot call va_start() twice on t… (#1162)
* Fix xclLogMsg va_list processing -- cannot call va_start() twice on the same va_list

Renamed xclhal2.h to xrt.h with the former being legacy wrapper. Moved experimental XRT APIs to xrt-next.h under experimental directory

Do not document now deprecated register read/write APIs in XMA

* Incorporate feedback on PR

* Minor update, use pre-increment
