+++ src/runtime_src/doc/toc/xrt.main.rst
@@ -1,4 +1,4 @@
 Xilinx Runtime (XRT) Core Library
 ---------------------------------
 
-.. include:: ../core/xclhal2.rst
+.. include:: ../core/xrt.rst
