+++ src/runtime_src/driver/include/experimental/CMakeLists.txt
@@ -1,0 +1,10 @@
+set(XRT_EXPERIMENTAL_HEADER_SRC
+  xrt-next.h)
+
+
+install (FILES ${XRT_EXPERIMENTAL_HEADER_SRC} DESTINATION ${XRT_INSTALL_DIR}/include/experimental)
+
+message("-- XRT experimental header files")
+foreach (header ${XRT_EXPERIMENTAL_HEADER_SRC})
+  message("-- ${header}")
+endforeach()
