+++ src/xma/include/xmaplugin.h
@@ -138,7 +138,7 @@ int32_t xma_plg_buffer_read(XmaHwSession     s_handle,
                             size_t           offset);
 
 /**
- *  
+ *
  * xma_plg_register_write() - Write kernel register(s)
  *
  *  This function writes the data provided and sets the specified AXI_Lite
@@ -164,34 +164,34 @@ int32_t xma_plg_register_prep_write(XmaHwSession     s_handle,
                                     size_t           offset);
 
 /**
- * xma_plg_schedule_work_item() - This function schedules a request to the XRT 
- * scheduler for execution of a kernel based on the saved state of the kernel registers 
- * supplied by the xma_plg_register_prep_write() function call.  The prep_write() keeps a 
- * shadow register map so that the schedule_work_item() can gather all registers 
+ * xma_plg_schedule_work_item() - This function schedules a request to the XRT
+ * scheduler for execution of a kernel based on the saved state of the kernel registers
+ * supplied by the xma_plg_register_prep_write() function call.  The prep_write() keeps a
+ * shadow register map so that the schedule_work_item() can gather all registers
  * and push a new work item onto the scheduler queue.  Work items are processed
  * in FIFO order.  After calling schedule_work_item() one or more times, the caller
- * can invoke xma_plg_is_work_item_done() to wait for one item of work to complete. 
+ * can invoke xma_plg_is_work_item_done() to wait for one item of work to complete.
  *
  * @s_handle: The session handle associated with this plugin instance
  *
  * RETURN:     XMA_SUCCESS on success
- *          
+ *
  * XMA_ERROR on failure
  *
  */
 int32_t xma_plg_schedule_work_item(XmaHwSession s_handle);
 
 /**
- * xma_plg_is_work_item_done() - This function checks if at least one work item 
- * previously submitted via xma_plg_schedule_work_item() has completed.  If the 
- * supplied timeout expires before a work item has completed, this function 
- * returns an error.  
+ * xma_plg_is_work_item_done() - This function checks if at least one work item
+ * previously submitted via xma_plg_schedule_work_item() has completed.  If the
+ * supplied timeout expires before a work item has completed, this function
+ * returns an error.
  *
  * @s_handle:      The session handle associated with this plugin instance
- * @timeout_in_ms: A timeout value in milliseconds 
+ * @timeout_in_ms: A timeout value in milliseconds
  *
  * RETURN:         XMA_SUCCESS on success
- *          
+ *
  * XMA_ERROR on timeout
  *
  */
@@ -201,8 +201,8 @@ void xma_plg_kernel_lock(XmaHwSession s_handle);
 void xma_plg_kernel_unlock(XmaHwSession s_handle);
 
 /**
- *  xma_plg_register_prep_write() - This function writes the data provided and sets 
- * the specified AXI_Lite register(s) exposed by a kernel. The base offset of 0 
+ *  xma_plg_register_prep_write() - This function writes the data provided and sets
+ * the specified AXI_Lite register(s) exposed by a kernel. The base offset of 0
  * is the beginning of the kernels AXI_Lite memory map as this function adds the required
  * offsets internally for the kernel and PCIe.
  *
@@ -213,7 +213,7 @@ void xma_plg_kernel_unlock(XmaHwSession s_handle);
  *                   register map
  *
  *  RETURN:          >=0 number of bytes written
- *           
+ *
  * <0 on failure
  *
  */
@@ -229,9 +229,10 @@ void xma_plg_kernel_wait_on_finish(XmaHwSession s_handle);
 void xma_plg_kernel_start(XmaHwSession s_handle);
 int32_t xma_plg_kernel_exec(XmaHwSession s_handle, bool wait_on_kernel_finish);
 
-/**
- *  xma_plg_register_write() - This function writes the data provided and sets the 
- * specified AXI_Lite register(s) exposed by a kernel. The base offset of 0 is the 
+/*
+ * Deprecated API
+ *  xma_plg_register_write() - This function writes the data provided and sets the
+ * specified AXI_Lite register(s) exposed by a kernel. The base offset of 0 is the
  * beginning of the kernels AXI_Lite memory map as this function adds the required
  *  offsets internally for the kernel and PCIe.
  *
@@ -242,7 +243,7 @@ int32_t xma_plg_kernel_exec(XmaHwSession s_handle, bool wait_on_kernel_finish);
  *                   register map
  *
  *  RETURN:          >=0 number of bytes written
- *           
+ *
  * <0 on failure
  *
  */
@@ -251,8 +252,8 @@ int32_t xma_plg_register_write(XmaHwSession     s_handle,
                                size_t           size,
                                size_t           offset) __attribute__ ((deprecated));
 
-/**
- *  
+/*
+ * Deprecated API
  * xma_plg_register_read() - Read kernel registers
  *
  *  This function reads the register(s) exposed by the kernel. The base offset
@@ -266,7 +267,7 @@ int32_t xma_plg_register_write(XmaHwSession     s_handle,
  *                   map
  *
  *  RETURN:     >=0 number of bytes read
- * 
+ *
  * <0 on failure
  *
  */
@@ -276,7 +277,7 @@ int32_t xma_plg_register_read(XmaHwSession     s_handle,
                               size_t           offset) __attribute__ ((deprecated));
 
 /**
- *  
+ *
  * xma_plg_register_dump() - Dump kernel registers
  *
  *  This function dumps the registers for a kernel up to the number of words
