+++ tests/message-mgmt/Makefile
@@ -1,6 +1,6 @@
 CC    = g++
 CFLAGS = -DDSA64 -g -std=c++11 mssg.cpp -I../../build/Debug/opt/xilinx/xrt/include
-LDFLAGS = -L../../build/Debug/opt/xilinx/xrt/lib -lxrt_core -lxilinxopencl -lpthread -lrt
+LDFLAGS = -L../../build/Debug/opt/xilinx/xrt/lib -lxrt_core -lpthread -lrt
 run:
 	@$(CC) $(CFLAGS) -o host.exe $(LDFLAGS)
